Switch from auto-detect swagger version to hardcoded version
